feat: refactor validators using zod
